\name{mygene}
\alias{mygene}
\alias{MyGene.info}
\docType{package}
\title{
Access MyGene.info services

}
\description{
MyGene.Info_ provides simple-to-use REST web services to query/retrieve gene annotation data. It's designed with simplicity and performance emphasized. *mygene* is an easy-to-use R wrapper to access MyGene.info services.
}
\details{
\tabular{ll}{
Package: \tab mygene\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2014-04-18\cr
License: \tab BSD\cr
Depends: \tab httr jsonlite\cr
}
}
\author{
Adam Mark, Chunlei Wu

Maintainer: Chunlei Wu <help@mygene.info>
 
}
\references{
Wu C, MacLeod I, Su AI (2013) BioGPS and MyGene.info: organizing online, gene-centric information. Nucl. Acids Res. 41(D1): D561-D565. 
http://mygene.info/doc/annotation_service.html
http://mygene.info/doc/query_service.html
}

\keyword{ package }
