\name{metadata}
\docType{methods}
\alias{metadata-methods}
\alias{metadata,MyGene-method}

\title{
metadata
}
\description{
Get metadata for MyGene.info services.
}
\usage{
metadata(x, ...)
}

\references{
http://mygene.info/v2/metadata
}

\examples{
## Get metadata 
## mygene<-MyGene()
## metadata(mygene)

## get available fields to return
## metadata(mygene)$available_fields
}

